Adding explicit bool->uint64_t conversion to PackedRow.h to please the VC compiler
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cryptominisat/branches/gaussian_differentmatrix@696 6c43cb19-1bce-4953-9960-ac6414807ffa

+++ Solver/PackedRow.h
@@ -141,7 +141,7 @@ public:
 
     inline void invert_is_true(const bool b = true)
     {
-        is_true_internal ^= b;
+        is_true_internal ^= (uint64_t)b;
     }
 
     inline void setBit(const uint i)
